payflow/requirements.txt
apimatic-core==0.2.20
apimatic-core-interfaces==0.1.6
apimatic-requests-client-adapter==0.1.7
asgiref==3.8.1
CacheControl==0.12.14
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jsonpickle==3.3.0
jsonpointer==2.4
msgpack==1.1.0
packaging==25.0
paypal-server-sdk==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
